requirements.txt
flask==2.0
Flask-SocketIO==5.0.2
pyyaml==5.4.1
